Remove PR tab mock fixtures
