+++ packages/coc/src/server/spa/client/react/features/pull-requests/PrQueueFilters.tsx
@@ -9,7 +9,7 @@ import {
     getQueueFilterDefinitions,
     type QueueFilter,
     type QueueFilterCounts,
-} from './pr-mock-data';
+} from './pr-derived-data';
 
 interface PrQueueFiltersProps {
     active: QueueFilter;
